Make highlightable line edit its own widget - QgsHighlightableLineEdit
