+++ python/gui/auto_generated/qgshighlightablelineedit.sip.in
@@ -1,0 +1,54 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/qgshighlightablelineedit.h                                   *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+class QgsHighlightableLineEdit: QgsFilterLineEdit
+{
+%Docstring
+
+A QgsFilterLineEdit subclass with the ability to "highlight" the edges of the widget.
+
+.. versionadded:: 3.14
+%End
+
+%TypeHeaderCode
+#include "qgshighlightablelineedit.h"
+%End
+  public:
+    QgsHighlightableLineEdit( QWidget *parent /TransferThis/ = 0 );
+
+    bool isHighlighted() const;
+%Docstring
+Returns ``True`` if the line edit is currently highlighted.
+
+.. seealso:: :py:func:`setHighlighted`
+%End
+
+    void setHighlighted( bool highlighted );
+%Docstring
+Sets whether the line edit is currently ``highlighted``.
+
+.. seealso:: :py:func:`isHighlighted`
+%End
+
+  protected:
+    virtual void paintEvent( QPaintEvent *e );
+
+
+};
+
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/qgshighlightablelineedit.h                                   *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
