+++ src/gui/qgsfilewidget.h
@@ -26,7 +26,7 @@ class QHBoxLayout;
 
 #include "qgis_gui.h"
 #include "qgis_sip.h"
-#include "qgsfilterlineedit.h"
+#include "qgshighlightablelineedit.h"
 
 /**
  * \ingroup gui
@@ -243,7 +243,7 @@ class GUI_EXPORT QgsFileWidget : public QWidget
  * or directories only. By default, dropping is limited to files only.
  * \note not available in Python bindings
  */
-class GUI_EXPORT QgsFileDropEdit: public QgsFilterLineEdit
+class GUI_EXPORT QgsFileDropEdit: public QgsHighlightableLineEdit
 {
     Q_OBJECT
 
@@ -259,7 +259,6 @@ class GUI_EXPORT QgsFileDropEdit: public QgsFilterLineEdit
     void dragEnterEvent( QDragEnterEvent *event ) override;
     void dragLeaveEvent( QDragLeaveEvent *event ) override;
     void dropEvent( QDropEvent *event ) override;
-    void paintEvent( QPaintEvent *e ) override;
 
   private:
 
@@ -268,7 +267,6 @@ class GUI_EXPORT QgsFileDropEdit: public QgsFilterLineEdit
 
     QStringList mAcceptableExtensions;
     QgsFileWidget::StorageMode mStorageMode = QgsFileWidget::GetFile;
-    bool mDragActive;
     friend class TestQgsFileWidget;
 };
